# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Chesed Train API",
    "version": "1.0.0",
    "description": "Core API surface for Chesed Train."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "TaskSlot": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "trainId": { "type": "string" },
          "date": { "type": "string", "format": "date-time" },
          "taskType": { "type": "string" },
          "status": { "type": "string" }
        }
      },
...
      "MealTrain": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "slug": { "type": "string" },
          "title": { "type": "string" },
          "recipientName": { "type": "string" },
          "isPublic": { "type": "boolean" },
          "taskSlots": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TaskSlot" }
          }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" },
          "pages": { "type": "integer" }
        }
      }
    }
  },
  "paths": {
...
    "/api/users/trains": {
      "get": {
        "summary": "List trains organized by the current user",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "User trains",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/MealTrain" }
                    },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/users/dashboard-stats": {
      "get": {
        "summary": "Dashboard stats for the current user",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Dashboard stats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "totalTrains": { "type": "integer" },
                    "activeTrains": { "type": "integer" },
                    "totalContributions": { "type": "integer" },
                    "totalDonationsAmount": { "type": "number" },
                    "upcomingDeliveries": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/TaskSlot" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}